requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
opencv-python==4.13.0.92
face-recognition==1.3.0
pandas==3.0.1
Pillow==12.1.1
